Move CircleCutout to its own file
